# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for E-Document Management System",
        "title": "E-Document API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@edocument.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5001",
    "basePath": "/api",
    "paths": {
...
        "/v1/documents": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Get all documents",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by document name",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/e-document-backend_internal_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/e-document-backend_internal_util.PaginatedData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new document",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Create document",
                "parameters": [
                    {
                        "description": "Document data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_domain.CreateDocumentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/e-document-backend_internal_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/e-document-backend_internal_domain.DocumentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            }
        },
        "/v1/documents/folder/{folderId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all documents in a specific folder",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Get documents by folder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder ID",
                        "name": "folderId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/e-document-backend_internal_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/e-document-backend_internal_domain.DocumentResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            }
        },
        "/v1/documents/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get document details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Get document by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/e-document-backend_internal_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/e-document-backend_internal_domain.DocumentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update document details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Update document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Document data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_domain.UpdateDocumentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/e-document-backend_internal_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/e-document-backend_internal_domain.DocumentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete document by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Delete document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            }
        },
        "/v1/documents/{id}/send": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send document to a receiver (user)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Send document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Send data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_domain.SendDocumentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            }
        },
        "/v1/documents/{id}/versions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all versions of a document",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Get document versions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/e-document-backend_internal_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/e-document-backend_internal_domain.VersionResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "e-document-backend_internal_domain.CreateDocumentRequest": {
            "type": "object",
            "required": [
                "doc_name"
            ],
            "properties": {
                "department_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "doc_name": {
                    "description": "Added form tag",
                    "type": "string"
                },
                "doc_path": {
                    "description": "Not from form directly, set by handler",
                    "type": "string"
                },
                "doc_type_id": {
                    "type": "string"
                },
                "folder_id": {
                    "type": "string"
                },
                "send_to_director": {
                    "type": "boolean"
                }
            }
        },
...
        "e-document-backend_internal_domain.DocumentResponse": {
            "type": "object",
            "properties": {
                "approval_status": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/e-document-backend_internal_domain.DepartmentApprovalStatus"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "department_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "doc_name": {
                    "type": "string"
                },
                "doc_no": {
                    "type": "string"
                },
                "doc_path": {
                    "type": "string"
                },
                "doc_type_id": {
                    "type": "string"
                },
                "doc_type_name": {
                    "type": "string"
                },
                "folder_id": {
                    "type": "string"
                },
                "folder_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "registrant_email": {
                    "type": "string"
                },
                "registrant_id": {
                    "type": "string"
                },
                "registrant_name": {
                    "type": "string"
                },
                "sector_name": {
                    "type": "string"
                },
                "send_to_director": {
                    "type": "boolean"
                },
                "status": {
                    "$ref": "#/definitions/e-document-backend_internal_domain.DocumentStatus"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "version_number": {
                    "type": "integer"
                }
            }
        },
...
        "e-document-backend_internal_domain.SendDocumentRequest": {
            "type": "object",
            "properties": {
                "receiver_id": {
                    "description": "Can be User UUID or Department ID (if implementing logic for that)",
                    "type": "string"
                },
                "receiver_type": {
                    "description": "\"user\" or \"department\"",
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                }
            }
        },
...
        "e-document-backend_internal_domain.UpdateDocumentRequest": {
            "type": "object",
            "properties": {
                "department_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "doc_name": {
                    "type": "string"
                },
                "doc_path": {
                    "description": "Not from form directly, set by handler",
                    "type": "string"
                },
                "doc_type_id": {
                    "type": "string"
                },
                "folder_id": {
                    "type": "string"
                },
                "send_to_director": {
                    "type": "boolean"
                }
            }
        },
...
        "e-document-backend_internal_domain.VersionResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "doc_id": {
                    "type": "string"
                },
                "doc_path": {
                    "type": "string"
                },
                "folder_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "version_number": {
                    "type": "integer"
                }
            }
        },
...
        "e-document-backend_internal_util.PaginatedData": {
            "type": "object",
            "properties": {
                "items": {}
            }
        },
...
        "e-document-backend_internal_util.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error_code": {
                    "$ref": "#/definitions/e-document-backend_internal_util.ErrorCode"
                },
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/e-document-backend_internal_util.PaginationInfo"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}